requirements.txt
line-bot-sdk
gunicorn
flask
boto3